move "primitive" db-sync types to a separate crate
